pyproject.toml
[tool.poetry]
name = "stock-django-be"
version = "0.1.0"
description = ""
authors = ["Jerome-Spirinova <jerome@spirinova.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
